+++ src/lib/directives/index.ts
@@ -1,0 +1,1 @@
+export { StInputAdjustable } from './st-input-adjustable';