+++ src/lib/st-page-title/st-page-title.component.spec.ts
@@ -1,6 +1,7 @@
 import { DebugElement } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
+import { FormsModule } from '@angular/forms';
 
 // Component
 import { StButtonModule } from '../st-button/st-button.module';
@@ -8,41 +9,32 @@ import { StPageTitleComponent } from './st-page-title.component';
 
 let component: StPageTitleComponent;
 let fixture: ComponentFixture<StPageTitleComponent>;
-
+let compiled: any;
 
 describe('StPageTitleComponent', () => {
-
-   beforeEach(async(() => {
-      TestBed.configureTestingModule({
-         imports: [StButtonModule],
-         declarations: [StPageTitleComponent]
+   beforeEach(
+      async(() => {
+         TestBed.configureTestingModule({
+            imports: [StButtonModule, FormsModule],
+            declarations: [StPageTitleComponent]
+         }).compileComponents(); // compile template and css
       })
-         .compileComponents();  // compile template and css
-   }));
+   );
 
    beforeEach(() => {
       fixture = TestBed.createComponent(StPageTitleComponent);
       component = fixture.componentInstance;
+      compiled = fixture.debugElement.nativeElement;
       component.qaTag = 'test qaTag';
    });
-
-   it('Throw error if not qaTag', () => {
-      component.leftButton = undefined;
-      component.qaTag = undefined;
-
-      try {
-         expect(fixture.detectChanges()).toThrow();
-      } catch (error) {
-         expect(error.message).toContain('ST-PAGE-TITLE: qa field is required');
-      }
-   });
-
 
    it(`Only when there is the input 'leftButton', button is shown before title`, () => {
       component.leftButton = undefined;
       fixture.detectChanges();
 
-      let button: HTMLButtonElement = fixture.nativeElement.querySelector('.st-page-title button');
+      let button: HTMLButtonElement = fixture.nativeElement.querySelector(
+         '.st-page-title button'
+      );
       expect(button).toBeNull();
 
       component.leftButton = 'icon-reply';
@@ -51,7 +43,6 @@ describe('StPageTitleComponent', () => {
       button = fixture.nativeElement.querySelector('.st-page-title button');
       expect(button).toBeDefined();
    });
-
 
    it('Emit when click on button', () => {
       let responseFunction = jasmine.createSpy('response');
@@ -68,5 +59,50 @@ describe('StPageTitleComponent', () => {
       fixture.detectChanges();
 
       expect(responseFunction).toHaveBeenCalled();
+   });
+
+   describe('When component is editable', () => {
+      it('Render input when it is a page title editable', () => {
+         expect(compiled.querySelector('input')).toBeNull();
+
+         component.editable = true;
+         fixture.detectChanges();
+
+         expect(compiled.querySelector('input')).toBeDefined();
+      });
+
+      it('Should be the focus on the input when clicking on the title', () => {
+         component.editable = true;
+         fixture.detectChanges();
+         component.onClickEdit();
+         fixture.detectChanges();
+
+         expect(fixture.componentInstance.focus).toBeTruthy();
+      });
+
+      it('Focus should be removed when we exit input', () => {
+         component.editable = true;
+         fixture.detectChanges();
+
+         let input = fixture.debugElement.query(By.css('input')).nativeElement;
+         input.dispatchEvent(new Event('focus'));
+         fixture.detectChanges();
+
+         input.dispatchEvent(new Event('blur'));
+         expect(fixture.componentInstance.focus).toBeFalsy();
+      });
+
+      it('Should change the value of title as input value', () => {
+         component.editable = true;
+         fixture.detectChanges();
+
+         let input = fixture.debugElement.query(By.css('input')).nativeElement;
+
+         input.value = 'test';
+         input.dispatchEvent(new Event('input'));
+         fixture.detectChanges();
+
+         expect(fixture.componentInstance.title).toEqual('test');
+      });
    });
 });