+++ src/lib/st-page-title/st-page-title.module.ts
@@ -1,12 +1,15 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 
 import { StButtonModule } from '../st-button/st-button.module';
 import { StPageTitleComponent } from './st-page-title.component';
 
+import { StInputAdjustable } from '../directives';
+
 @NgModule({
-   imports: [CommonModule, StButtonModule],
-   declarations: [StPageTitleComponent],
+   imports: [CommonModule, FormsModule, StButtonModule],
+   declarations: [StPageTitleComponent, StInputAdjustable],
    exports: [StPageTitleComponent]
 })
-export class StPageTitleModule { }
+export class StPageTitleModule {}